+++ packages/sui-polyfills/src/index.js
@@ -21,6 +21,7 @@ require('core-js/features/string/starts-with')
 require('core-js/features/string/trim')
 require('core-js/features/string/pad-end')
 require('core-js/features/string/pad-start')
+require('core-js/features/string/match-all')
 
 require('core-js/features/url')
 require('core-js/features/url-search-params')